requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-split-settings==1.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
Markdown==3.4.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==22.0
pathspec==0.10.3
platformdirs==2.6.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pymyob==1.2.19
pyrsistent==0.19.2
pytz==2022.7
PyYAML==6.0
redis==4.4.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
swagger-spec-validator==3.0.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
